+++ src/orm/utils/string.cpp
@@ -16,8 +16,8 @@ namespace Orm::Utils::String
 QString toSnake(QString string)
 {
     // RegExp not used for performance reasons
-    std::vector<unsigned int> positions;
-    positions.reserve(string.size() / 2 + 2);
+    std::vector<int> positions;
+    positions.reserve(static_cast<std::size_t>(string.size() / 2 + 2));
 
     for (auto i = 0; i < string.size(); ++i) {
         const auto ch = string.at(i);
